Replace deprecated NSNotificationCentre with UNNotificationCentre
Sadly this requires the apps to be signed, so it only really works for store-distributed apps.
But it's required because of the deprecations.

Fixes #1833
