+++ app/app_darwin.go
@@ -1,75 +1,39 @@
 // +build !ci
 
-// +build !ios
-
 package app
 
 /*
 #cgo CFLAGS: -x objective-c
-#cgo LDFLAGS: -framework Foundation
+#cgo LDFLAGS: -framework Foundation -framework UserNotifications
 
-#include <AppKit/AppKit.h>
+#include <stdbool.h>
+#include <stdlib.h>
 
 bool isBundled();
-bool isDarkMode();
-void sendNotification(const char *title, const char *content);
-void watchTheme();
+void sendNotification(char *title, char *content);
 */
 import "C"
 import (
 	"fmt"
-	"net/url"
-	"os"
-	"path/filepath"
 	"strings"
 	"unsafe"
 
 	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/theme"
-
 	"golang.org/x/sys/execabs"
 )
 
-func defaultVariant() fyne.ThemeVariant {
-	if C.isDarkMode() {
-		return theme.VariantDark
-	}
-	return theme.VariantLight
-}
-
-func rootConfigDir() string {
-	homeDir, _ := os.UserHomeDir()
-
-	desktopConfig := filepath.Join(filepath.Join(homeDir, "Library"), "Preferences")
-	return filepath.Join(desktopConfig, "fyne")
-}
-
-func (a *fyneApp) OpenURL(url *url.URL) error {
-	cmd := a.exec("open", url.String())
-	cmd.Stdin, cmd.Stdout, cmd.Stderr = os.Stdin, os.Stdout, os.Stderr
-	return cmd.Run()
-}
-
 func (a *fyneApp) SendNotification(n *fyne.Notification) {
 	if C.isBundled() {
-		title := C.CString(n.Title)
-		defer C.free(unsafe.Pointer(title))
-		content := C.CString(n.Content)
-		defer C.free(unsafe.Pointer(content))
+		titleStr := C.CString(n.Title)
+		defer C.free(unsafe.Pointer(titleStr))
+		contentStr := C.CString(n.Content)
+		defer C.free(unsafe.Pointer(contentStr))
 
-		C.sendNotification(title, content)
+		C.sendNotification(titleStr, contentStr)
 		return
 	}
 
-	title := escapeNotificationString(n.Title)
-	content := escapeNotificationString(n.Content)
-	template := `display notification "%s" with title "%s"`
-	script := fmt.Sprintf(template, content, title)
-
-	err := execabs.Command("osascript", "-e", script).Start()
-	if err != nil {
-		fyne.LogError("Failed to launch darwin notify script", err)
-	}
+	fallbackNotification(n.Title, n.Content)
 }
 
 func escapeNotificationString(in string) string {
@@ -77,11 +41,19 @@ func escapeNotificationString(in string) string {
 	return strings.ReplaceAll(noSlash, "\"", "\\\"")
 }
 
-//export themeChanged
-func themeChanged() {
-	fyne.CurrentApp().Settings().(*settings).setupTheme()
+//export fallbackSend
+func fallbackSend(cTitle, cContent *C.char) {
+	title := C.GoString(cTitle)
+	content := C.GoString(cContent)
+	fallbackNotification(title, content)
 }
 
-func watchTheme() {
-	C.watchTheme()
+func fallbackNotification(title, content string) {
+	template := `display notification "%s" with title "%s"`
+	script := fmt.Sprintf(template, escapeNotificationString(content), escapeNotificationString(title))
+
+	err := execabs.Command("osascript", "-e", script).Start()
+	if err != nil {
+		fyne.LogError("Failed to launch darwin notify script", err)
+	}
 }
